inline addHealthCheck, prepare for removal
